requirements.txt
pandas
numpy
selenium
beautifulsoup4
webdriver-manager
apache-airflow-providers-microsoft-azure